requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
marshmallow==3.20.1
bcrypt==4.0.1
PyJWT==2.8.0
Werkzeug==2.3.7